Refactor PyTorch model fixtures into separate conftest.py
These fixtures are only used by pytorch to mdf conversion tests. Lets keep them separate.
